+++ app/Http/Controllers/AdminClient/ProductController.php
@@ -63,6 +63,9 @@ class ProductController extends Controller
     {
         $product = Product::find($id);
 
+        //controlo de que sea el dueño del producto
+        $this->authorize('passProduct', $product);
+
         $commerce = Commerce::where('user_id', auth()->user()->id)
             ->first();
 
@@ -75,6 +78,9 @@ class ProductController extends Controller
     {
 
         $product = Product::find($id);
+
+        //controlo de que sea el dueño del producto
+        $this->authorize('passProduct', $product);
 
         $product->fill($request->all())->save();
 
@@ -104,6 +110,8 @@ class ProductController extends Controller
     {
         $product = Product::find($id);
 
+        //controlo de que sea el dueño del producto
+        $this->authorize('passProduct', $product);
         $product->available = 'YES';
         $product->update();
 
@@ -115,6 +123,8 @@ class ProductController extends Controller
     {
         $product = Product::find($id);
 
+        //controlo de que sea el dueño del producto
+        $this->authorize('passProduct', $product);
         $product->available = 'NO';
         $product->update();
 
@@ -125,6 +135,9 @@ class ProductController extends Controller
     public function destroy($id)
     {
         $product = Product::find($id);
+
+        //controlo de que sea el dueño del producto
+        $this->authorize('passProduct', $product);
 
         if(File::exists(public_path('images/thumbnail/products/' . $product->photo))){
             File::delete(public_path('images/thumbnail/products/' . $product->photo));
